Mudança para estado global com context
