+++ .gitignore
@@ -1,3 +1,2 @@
-global
 build
 node_modules
