+++ src/global/Context.js
@@ -1,0 +1,5 @@
+import {createContext} from 'react'
+
+const Context = createContext()
+
+export default Context
